/simplify
Just a bit. Courtesy of Claude. I verified all simplifications.
