+++ packages/core/src/scouts/scoutPrompts.ts
@@ -1,0 +1,42 @@
+// Templated prompts behind the scout chat CTA chips. The agent leans on the
+// exploring-signals-scouts skill from the PostHog MCP.
+
+export const SCOUT_FLEET_OVERVIEW_PROMPT = `How is my scout fleet performing?
+
+Use the exploring-signals-scouts skill from the PostHog MCP to survey the signals scout fleet on this project and give me a high-level overview:
+
+- The fleet: which scouts exist, enabled vs disabled, and their cadences
+- Recent run health: success rate, failures and timeouts, anything stuck
+- Output: which scouts emitted signals recently, emit rate, signal-to-noise
+- Memory: notable scratchpad entries the fleet has learned
+- Recommendations: anything misconfigured, noisy, or worth tuning
+
+Lead with a short overall verdict, then per-scout notes only where something is notable. If the skill is unavailable, fall back to the signals-scout MCP tools directly (config list, runs list, scratchpad search).`;
+
+export const SCOUT_RECENT_SIGNALS_PROMPT = `What signals have my scouts emitted recently?
+
+Use the exploring-signals-scouts skill from the PostHog MCP to pull the most recent scout runs that emitted findings and walk me through the signals:
+
+- What each signal says, in plain language
+- Which scout emitted it, when, and its severity/confidence where available
+- Whether it looks genuinely actionable or like noise
+
+Group by scout, newest first. Close with a short note on overall signal quality and any scouts that look noisy or suspiciously silent. If the skill is unavailable, fall back to the signals-scout MCP tools directly (runs list with emitted filter, run emissions).`;
+
+/** Per-scout variant of the templated questions, scoped to one skill. */
+export function buildScoutCheckinPrompt(
+  skillName: string,
+  displayName: string,
+): string {
+  return `How is my ${displayName} scout performing?
+
+Use the exploring-signals-scouts skill from the PostHog MCP to dig into the \`${skillName}\` scout on this project:
+
+- Its config: enabled, cadence, dry-run posture
+- Recent run history: successes, failures, timeouts, durations
+- Signals it emitted recently and whether they look genuinely actionable
+- Scratchpad memory the fleet holds that relates to this scout
+- Whether its scope, thresholds, and schedule look right — suggest tuning if not
+
+Lead with a short verdict. If the skill is unavailable, fall back to the signals-scout MCP tools directly (config list, runs list, run emissions, scratchpad search).`;
+}
