+++ packages/ui/src/features/scouts/hooks/useScoutRunEmissions.ts
@@ -3,10 +3,7 @@ import { useAuthenticatedQuery } from "@posthog/ui/hooks/useAuthenticatedQuery";
 import { useAuthStateValue } from "../../auth/store";
 import { scoutQueryKeys } from "./scoutQueryKeys";
 
-export function useScoutRunEmissions(
-  runId: string,
-  options?: { enabled?: boolean },
-) {
+export function useScoutRunEmissions(runId: string) {
   const projectId = useAuthStateValue((state) => state.currentProjectId);
   return useAuthenticatedQuery<ScoutEmission[]>(
     scoutQueryKeys.emissions(projectId, runId),
@@ -15,7 +12,7 @@ export function useScoutRunEmissions(
         ? client.listScoutRunEmissions(projectId, runId)
         : Promise.resolve([]),
     {
-      enabled: !!projectId && !!runId && (options?.enabled ?? true),
+      enabled: !!projectId && !!runId,
       staleTime: 60_000,
     },
   );
